requirements.txt
# ── Core API ──────────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6   # Required for FastAPI Form/File uploads
python-dotenv>=1.0.0

# ── LLM (Groq via OpenAI-compatible SDK) ──────────────────────────────
openai>=1.0.0             # Groq uses the OpenAI SDK with a custom base_url

# ── Document loading (lightweight, no torch/faiss required) ───────────
pypdf>=3.0.0              # PDF support
python-docx>=1.0.0        # DOCX support
pillow>=1.0.0
